requirements.txt
discord.py==0.16.12
Pillow==5.3.0